Add check: Incorrect delimiter (#20)
* Create incorrect_delimiter file scaffold

* Implement IncorrectDelimiterChecker

* Include IncorrectDelimiterChecker in checklist()

Change checklist function to return a Boxed trait object so that
IncorrectDelimiterChecker can be included.

* Reformat using rustfmt with `max-width` at 80

Previously it was set at 100 on this file. Now changing to be in line
with the rest of the project.

* Add IncorrectDelimiter to README

* Add default Warning for IncorrectDelimiterChecker

Add a warning field to the struct and create an implementaion of
```default()```. The returned Warning in ```IncorrectDelimiter.check()```
will use ```String.replace()``` as the ```format!``` macro only works on
string literals.

* Update checklist() to use default() for linter

* Update README with IncorrectDelimiter example

* Remove unnecessary whitespace check in linter

In ```IncorrectDelimiterChecker.check()```:

- Remove early return if line starts with " "
- Trim whitespace of ```key``` before checking for delimiters
- Add tests to make sure the linter does not return non-delimiter
  whitespace errors

* FIx typo in README

Co-Authored-By: Grachev Mikhail <work@mgrachev.com>