Fix #103: Remove the unwrap method while working with current directory
current_dir.to_str is replaced with as_os_str which yields the OsString.
The argument type of get_args() function is changed from &str to &OsStr.

The clap code to get the argument `input` was using default_value method
which was converting OsString to &str. Instead default_value method is
swapped with default_value_os which accepts &OsString