Create helper types and methods for CLI testing
Add the TestDir type which can create, operate on, and remove
temporary directories. Also, it provides an easy way to create a
TestFile and subdirectories within the TestDir. These methods make it
easy to setup and cleanup  Command Line Interface tests.

Moreover, TestDir has methods to check Command success and failure.
This is done by wrapping the "assert_cmd" and "tempfile" crates.