Refactor Check unit tests (#421)
* refactor: add testing macro for check trait impls

check_tester macro helps to remove the boiler plate when testing the
impls of the Check trait.

* docs: add documentation for check_tester macro

Internal documentation for the check_tester macro to make it
easier to understand and use with some examples.

* chore: rust fmt

* refactor: update new check to use check_tester

Refactors SubstitutionKeyChecker unit tests to use the check_tester
macro inline with the other Check impls.

* refactor: replace test macro for helper function

Replaces the check_tester! declarative macro with a generic test
helper function. This will be more explicit and easier to update
than the check_tester macro.

* chore: rust fmt

* chore: update changelog

Updates the changelog to include the changes made in the PR #421.

* fix: solves clippy warning

Fixes the 'inconsistent-struct-constructor' clippy warning.