feat: add check for values that require surrounding quotes
When evaluating variables, values with spaces will only evaluate the
first word/character before a space and ignore the rest. This is
especially bad because this fails silently in most cases and can lead
to confusing scenarios.

This ensures that any value that contains a space must have quotes
surrounding it so that it is valid.

This does not consider any value that has a leading or trailing space
since that is managed by other checks/fixes.

Further Fixes: https://github.com/dotenv-linter/dotenv-linter/issues/343